Update rlwrap
By using a different tarball from the same site that had a configure file, was able to avoid patching configure.ac and generating a configure script, thus reducing dependencies and build time.  Also used `CPPFLAGS` instead of `CFLAGS` for specifying the location of some header files and this seemed to make configure happier.